refactor(engineer_loop/verification): split tests to new sibling tests_verification_extra.rs (#1266)
Inline test module in verification.rs is moved into NEW sibling
tests_verification_extra.rs. Existing tests_verification.rs (759 LOC)
was not extended (already over 400 — separate cleanup needed).

- src/engineer_loop/verification.rs: 524 → 267 LOC (under 400-line rule)
- src/engineer_loop/tests_verification_extra.rs: new (~258 LOC)
- src/engineer_loop/mod.rs: register mod tests_verification_extra
- Imports prepended (matching tests_verification.rs surface)

Pure refactor: zero logic changes, all 16 tests pass.

Closes 19/70 of #1266 (this loop iteration), 28 PRs cumulative since #1285.
